+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ CHANGELOG
 v1.03.15
 =========
 
+* Add the ability to delete a Customer from the Edit Customer admin page. This
+  will anonymous & move their order & related addresses to the
+  ``gardens+deleted@southernexposure.com`` account and then delete all other
+  data for the Customer.
 * Show Product shipping restrictions on the Cart & Product Details pages.
 * Fix spacing around general error messages on the Checkout page.
 * Add a new "Cannot Ship To" field to the Add/Edit Product pages. This is an